Move subscriptions to solana-rpc
