+++ core/Cargo.toml
@@ -33,8 +33,6 @@ jsonrpc-core = "17.0.0"
 jsonrpc-core-client = { version = "17.0.0", features = ["ipc", "ws"] }
 jsonrpc-derive = "17.0.0"
 jsonrpc-http-server = "17.0.0"
-jsonrpc-pubsub = "17.0.0"
-jsonrpc-ws-server = "17.0.0"
 libc = "0.2.81"
 log = "0.4.11"
 lru = "0.6.1"
