+++ core/src/completed_data_sets_service.rs
@@ -1,8 +1,7 @@
-use crate::rpc_subscriptions::RpcSubscriptions;
 use crossbeam_channel::{Receiver, RecvTimeoutError, Sender};
 use solana_ledger::blockstore::{Blockstore, CompletedDataSetInfo};
 use solana_ledger::entry::Entry;
-use solana_rpc::max_slots::MaxSlots;
+use solana_rpc::{max_slots::MaxSlots, rpc_subscriptions::RpcSubscriptions};
 use solana_sdk::signature::Signature;
 use std::{
     sync::{
