+++ core/src/lib.rs
@@ -44,7 +44,6 @@ pub mod heaviest_subtree_fork_choice;
 pub mod latest_validator_votes_for_frozen_banks;
 pub mod ledger_cleanup_service;
 pub mod optimistic_confirmation_verifier;
-pub mod optimistically_confirmed_bank_tracker;
 pub mod outstanding_requests;
 pub mod packet_hasher;
 pub mod ping_pong;
@@ -61,12 +60,8 @@ mod result;
 pub mod retransmit_stage;
 pub mod rewards_recorder_service;
 pub mod rpc;
-pub mod rpc_completed_slots_service;
 pub mod rpc_health;
-pub mod rpc_pubsub;
-pub mod rpc_pubsub_service;
 pub mod rpc_service;
-pub mod rpc_subscriptions;
 pub mod send_transaction_service;
 pub mod serve_repair;
 pub mod serve_repair_service;
