+++ rpc/src/lib.rs
@@ -1,2 +1,21 @@
 #![allow(clippy::integer_arithmetic)]
 pub mod max_slots;
+pub mod optimistically_confirmed_bank_tracker;
+pub mod parsed_token_accounts;
+pub mod rpc_completed_slots_service;
+pub mod rpc_pubsub;
+pub mod rpc_pubsub_service;
+pub mod rpc_subscriptions;
+
+#[macro_use]
+extern crate log;
+
+#[macro_use]
+extern crate serde_derive;
+
+#[cfg(test)]
+#[macro_use]
+extern crate serde_json;
+
+#[macro_use]
+extern crate solana_metrics;
