+++ rpc/src/rpc_completed_slots_service.rs
