+++ rpc/src/rpc_pubsub.rs
@@ -542,11 +542,9 @@ impl RpcSolPubSub for RpcSolPubSubImpl {
 mod tests {
     use super::*;
     use crate::{
-        cluster_info_vote_listener::{ClusterInfoVoteListener, VoteTracker},
         optimistically_confirmed_bank_tracker::OptimisticallyConfirmedBank,
         rpc_subscriptions::tests::robust_poll_or_panic,
     };
-    use crossbeam_channel::unbounded;
     use jsonrpc_core::{futures::channel::mpsc, Response};
     use jsonrpc_pubsub::{PubSubHandler, Session};
     use serial_test::serial;
@@ -575,7 +573,7 @@ mod tests {
         self, stake_instruction,
         stake_state::{Authorized, Lockup, StakeAuthorize, StakeState},
     };
-    use solana_vote_program::vote_transaction;
+    use solana_vote_program::vote_state::Vote;
     use std::{
         sync::{atomic::AtomicBool, RwLock},
         thread::sleep,
@@ -1227,9 +1225,7 @@ mod tests {
         );
         let exit = Arc::new(AtomicBool::new(false));
         let bank = Bank::new(&genesis_config);
-        let bank_forks = BankForks::new(bank);
-        let bank = bank_forks.get(0).unwrap().clone();
-        let bank_forks = Arc::new(RwLock::new(bank_forks));
+        let bank_forks = Arc::new(RwLock::new(BankForks::new(bank)));
 
         // Setup RPC
         let mut rpc = RpcSolPubSubImpl::default_with_bank_forks(bank_forks.clone());
@@ -1239,52 +1235,22 @@ mod tests {
         // Setup Subscriptions
         let optimistically_confirmed_bank =
             OptimisticallyConfirmedBank::locked_from_bank_forks_root(&bank_forks);
-        let subscriptions = RpcSubscriptions::new_with_vote_subscription(
+        let subscriptions = Arc::new(RpcSubscriptions::new_with_vote_subscription(
             &exit,
             bank_forks,
             block_commitment_cache,
             optimistically_confirmed_bank,
             true,
-        );
-        rpc.subscriptions = Arc::new(subscriptions);
+        ));
+        rpc.subscriptions = subscriptions.clone();
         rpc.vote_subscribe(session, subscriber);
 
-        // Create some voters at genesis
-        let vote_tracker = VoteTracker::new(&bank);
-        let (votes_sender, votes_receiver) = unbounded();
-        let (vote_tracker, validator_voting_keypairs) =
-            (Arc::new(vote_tracker), validator_voting_keypairs);
-
-        let vote_slots = vec![1, 2];
-        validator_voting_keypairs.iter().for_each(|keypairs| {
-            let node_keypair = &keypairs.node_keypair;
-            let vote_keypair = &keypairs.vote_keypair;
-            let vote_tx = vote_transaction::new_vote_transaction(
-                vote_slots.clone(),
-                Hash::default(),
-                Hash::default(),
-                node_keypair,
-                vote_keypair,
-                vote_keypair,
-                None,
-            );
-            votes_sender.send(vec![vote_tx]).unwrap();
-        });
-
-        // Process votes and check they were notified.
-        let (verified_vote_sender, _verified_vote_receiver) = unbounded();
-        let (gossip_verified_vote_hash_sender, _gossip_verified_vote_hash_receiver) = unbounded();
-        let (_replay_votes_sender, replay_votes_receiver) = unbounded();
-        ClusterInfoVoteListener::get_and_process_votes_for_tests(
-            &votes_receiver,
-            &vote_tracker,
-            &bank,
-            &rpc.subscriptions,
-            &gossip_verified_vote_hash_sender,
-            &verified_vote_sender,
-            &replay_votes_receiver,
-        )
-        .unwrap();
+        let vote = Vote {
+            slots: vec![1, 2],
+            hash: Hash::default(),
+            timestamp: None,
+        };
+        subscriptions.notify_vote(&vote);
 
         let (response, _) = robust_poll_or_panic(receiver);
         assert_eq!(
