+++ validator/Cargo.toml
@@ -39,6 +39,7 @@ solana-logger = { path = "../logger", version = "=1.7.0" }
 solana-metrics = { path = "../metrics", version = "=1.7.0" }
 solana-net-utils = { path = "../net-utils", version = "=1.7.0" }
 solana-perf = { path = "../perf", version = "=1.7.0" }
+solana-rpc = { path = "../rpc", version = "=1.7.0" }
 solana-runtime = { path = "../runtime", version = "=1.7.0" }
 solana-sdk = { path = "../sdk", version = "=1.7.0" }
 solana-version = { path = "../version", version = "=1.7.0" }
